requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=2.2.2=pypi_0
ale-py=0.11.0=pypi_0
blas=1.0=openblas
bzip2=1.0.8=h80987f9_6
ca-certificates=2025.2.25=hca03da5_0
cloudpickle=3.0.0=py311hca03da5_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
deprecated=1.2.13=py311hca03da5_0
farama-notifications=0.0.4=py311hca03da5_0
filelock=3.17.0=py311hca03da5_0
fonttools=4.58.0=pypi_0
fsspec=2024.12.0=py311hca03da5_0
gmp=6.3.0=h313beb8_0
gmpy2=2.2.1=py311h5c1b81f_0
grpcio=1.71.0=pypi_0
gymnasium=1.1.1=pypi_0
importlib-metadata=8.5.0=py311hca03da5_0
jax-jumpy=1.0.0=py311hca03da5_0
jinja2=3.1.6=py311hca03da5_0
kiwisolver=1.4.8=pypi_0
libabseil=20250127.0=cxx17_h313beb8_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.29=hea593b9_0
libprotobuf=5.29.3=h9f9f828_0
libtorch=2.5.1=gpu_mps_h7b1bc93_204
libuv=1.48.0=h80987f9_0
llvm-openmp=14.0.6=hc6e5704_0
markdown=3.8=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py311h80987f9_0
matplotlib=3.10.3=pypi_0
mdurl=0.1.2=pypi_0
mpc=1.3.1=h80987f9_0
mpfr=4.2.1=h80987f9_0
mpmath=1.3.0=py311hca03da5_0
ncurses=6.4=h313beb8_0
networkx=3.4.2=py311hca03da5_0
nomkl=3.0=0
numpy=1.26.4=py311he598dae_0
numpy-base=1.26.4=py311hfbfe69c_0
opencv-python=4.11.0.86=pypi_0
openssl=3.0.16=h02f6b3c_0
opentelemetry-api=1.30.0=py311hca03da5_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
pillow=11.2.1=pypi_0
pip=25.1=pyhc872135_2
protobuf=6.30.2=pypi_0
psutil=7.0.0=pypi_0
pygame=2.6.1=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.3=pypi_0
python=3.11.11=hb885b13_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.5.1=gpu_mps_py311hbd1d365_204
pytz=2025.2=pypi_0
readline=8.2=h1a28f6b_0
rich=14.0.0=pypi_0
setuptools=78.1.1=py311hca03da5_0
six=1.17.0=pypi_0
sleef=3.5.1=h80987f9_2
sqlite=3.45.3=h80987f9_0
stable-baselines3=2.6.0=pypi_0
sympy=1.13.3=py311hca03da5_1
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tk=8.6.14=h6ba3021_0
tqdm=4.67.1=pypi_0
typing-extensions=4.12.2=py311hca03da5_0
typing_extensions=4.12.2=py311hca03da5_0
tzdata=2025.2=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py311hca03da5_0
wrapt=1.17.0=py311h80987f9_0
xz=5.6.4=h80987f9_1
yaml=0.2.5=h1a28f6b_0
zipp=3.21.0=py311hca03da5_0
zlib=1.2.13=h18a0788_1
